requirements.txt
pandas==2.2.2
pyarrow>=7.0,<14.0
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn==1.4.2
spacy==3.7.4
nltk==3.9.1
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
plotly==5.24.1
streamlit==1.38.0
tqdm>=4.67.0

# Big Data Tools (Academic Requirements)
pyspark==3.5.0
delta-spark==3.0.0
apache-airflow==2.8.0
kafka-python==2.0.2
mlflow==2.8.0
xgboost==2.0.0

# Additional ML and Analytics
joblib==1.3.2
matplotlib==3.8.0
seaborn==0.13.0
